[projectfirma/#1339] Calculate value of Technical Assistance as a PM Stub out the new pipeline where "GetReportedPerformanceMeasureValues" is delegated to PMDataSourceType
