+++ Source/ProjectFirma.Web/Models/PerformanceMeasureDataSourceType.cs
@@ -1,0 +1,36 @@
+using System.Collections.Generic;
+using System.Linq;
+using ProjectFirma.Web.Common;
+
+namespace ProjectFirma.Web.Models
+{
+    public abstract partial class PerformanceMeasureDataSourceType
+    {
+        public virtual List<PerformanceMeasureReportedValue> GetReportedPerformanceMeasureValues(PerformanceMeasure performanceMeasure,
+            List<int> projectIDs)
+        {
+            List<PerformanceMeasureActual> performanceMeasureActuals;
+            if (projectIDs == null || !projectIDs.Any())
+            {
+                performanceMeasureActuals = HttpRequestStorage.DatabaseEntities.PerformanceMeasureActuals.Where(pmav => pmav.PerformanceMeasureID == performanceMeasure.PerformanceMeasureID).ToList();
+            }
+            else
+            {
+                performanceMeasureActuals =
+                    HttpRequestStorage.DatabaseEntities.PerformanceMeasureActuals.Where(
+                        pmav => pmav.PerformanceMeasureID == performanceMeasure.PerformanceMeasureID && projectIDs.Contains(pmav.ProjectID)).ToList();
+            }
+            var performanceMeasureReportedValues = PerformanceMeasureReportedValue.MakeFromList(performanceMeasureActuals);
+            return performanceMeasureReportedValues.OrderByDescending(pma => pma.CalendarYear).ThenBy(pma => pma.ProjectName).ToList();
+        }
+    }
+
+    public partial class PerformanceMeasureDataSourceTypeTechnicalAssistanceValue
+    {
+        public override List<PerformanceMeasureReportedValue> GetReportedPerformanceMeasureValues(
+            PerformanceMeasure performanceMeasure, List<int> projectIDs)
+        {
+            return new List<PerformanceMeasureReportedValue>();
+        }
+    }
+}
