+++ Source/ProjectFirma.Web/ProjectFirma.Web.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="12.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -394,6 +394,7 @@
     <Compile Include="Models\Generated\WatershedPrimaryKey.Binding.cs" />
     <Compile Include="Models\IProjectCustomAttribute.cs" />
     <Compile Include="Models\IProjectCustomAttributeValue.cs" />
+    <Compile Include="Models\PerformanceMeasureDataSourceType.cs" />
     <Compile Include="Models\ProjectCustomAttributeDataType.cs" />
     <Compile Include="Models\ProjectCustomAttributeSimple.cs" />
     <Compile Include="Models\ProjectCustomAttributeType.cs" />
